I messed up the previous git repo (got into a detached head that I were unable to fix), so you're not getting proper commits this time. Not that anyone would check them.
If you're not doing it yet, you should be using git. It basically tracks the changes in your code and allows you to get back to a previous state if you mess something up (probability of it happening goes to 1 as size of the file increases).
Git is relatively easy to learn and you have to go out of your way to mess the repo up like I did. Also, it (or some other SVC) is required in pretty much any job, so if you want to get a job as a software dev, the sooner you start learning git, the more practice with it you will get and the better for your job interview.
tl;dr - git is gud, use it friend